requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
numpy==2.3.3
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
transformers
torch
pypdf
pymupdf
accelerate
tqdm
